Rename engine package to machine
