+++ app/machine/instructions/halt.go
@@ -1,6 +1,6 @@
 package instructions
 
-import "github/com/lucasbn/sqlite-clone/app/engine/state"
+import "github/com/lucasbn/sqlite-clone/app/machine/state"
 
 type Halt struct{}
