+++ app/machine/instructions/instruction.go
@@ -1,6 +1,6 @@
 package instructions
 
-import "github/com/lucasbn/sqlite-clone/app/engine/state"
+import "github/com/lucasbn/sqlite-clone/app/machine/state"
 
 type Instruction interface {
 	Execute(s state.MachineState) state.MachineState
