+++ app/machine/instructions/integer.go
@@ -1,6 +1,6 @@
 package instructions
 
-import "github/com/lucasbn/sqlite-clone/app/engine/state"
+import "github/com/lucasbn/sqlite-clone/app/machine/state"
 
 type Integer struct {
 	Register int
