+++ app/machine/registers/register.go
